Change data type of memory limit to uint64

+++ model/problem.go
@@ -68,7 +68,7 @@ type Problem struct {
 	Source       string     `json:"source"`
 	Difficulty   Difficulty `json:"difficulty"`
 	TimeLimit    float64    `json:"time_limit"`
-	MemoryLimit  float64    `json:"memory_limit"`
+	MemoryLimit  uint64     `json:"memory_limit"`
 	Description  string     `json:"description"`
 	Input        string     `json:"input"`
 	Output       string     `json:"output"`
